feat: migration.goにmigrationの処理を移行
