+++ backend/Dockerfile
@@ -27,4 +27,4 @@ COPY go.mod go.sum ./
 RUN go mod download
 EXPOSE 8000
 
-CMD ["go", "run", "main.go"]
+CMD ["go", "run", "main.go", "migration.go"]
